@@ -1,6 +1,6 @@
 {
   "name": "react-scrollbar",
-  "version": "0.5.1",
+  "version": "0.6.0",
   "description": "ScrollArea component for react",
   "main": "./dist/scrollArea.js",
   "scripts": {
@@ -23,8 +23,10 @@
   "dependencies": {
     "config": "^1.24.0",
     "line-height": "^0.1.1",
-    "react": "^15.4.1",
-    "react-motion": "^0.4.3"
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-motion": "^0.5.0"
   },
   "devDependencies": {
     "babel": "^6.5.2",
@@ -57,8 +59,6 @@
     "mocha": "^3.2.0",
     "object-assign": "^4.0.1",
     "phantomjs-prebuilt": "^2.1.3",
-    "react-addons-test-utils": "^15.1.0",
-    "react-dom": "^15.4.1",
     "run-sequence": "^1.1.4",
     "style-loader": "^0.13.0",
     "webpack": "^1.12.2",
